feat: widen Qrz coordinates from i16 to i32
Expands world address space from ~32K tiles (~3hrs walking) to ~2B tiles
per axis. All three Qrz fields (q, r, z) widened uniformly to avoid
mixed-type arithmetic. Mechanical type change only — no logic changes.

+++ crates/client/src/systems/tier_lock_range_indicator.rs
@@ -124,7 +124,7 @@ pub fn update(
     let mut tiles_in_range = Vec::new();
 
     // Search in a square bounding box (max_dist * 2 + 1) around player
-    let search_radius = max_dist as i16;
+    let search_radius = max_dist as i32;
     for dq in -search_radius..=search_radius {
         for dr in -search_radius..=search_radius {
             let tile_qrz = Qrz {
